pyproject.toml
[project]
name = "requests-docs-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.0.8",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-deepseek>=0.1.3",
    "langchain-huggingface>=0.1.2",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "text2vec>=1.3.5",
    "transformers[torch]>=4.51.3",
]
